requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
PyYAML
